requirements.txt
pinecone-client
sentence-transformers
streamlit
beautifulsoup4
requests